webserver/requirements.txt
Flask
mysql-connector-python
flask-mysqldb
pyyaml